requirements.txt
# Core framework
flask
flask-cors

# API clients  
openai
requests

# Document generation
python-pptx
python-docx
Pillow

# Authentication
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Database
psycopg2-binary

# Environment and configuration
python-dotenv

# Production server
gunicorn

# Testing
pytest

# Job Queue & Background Processing
celery[redis]==5.3.4
redis==4.6.0
kombu==5.3.4

# Email Services
flask-mail==0.9.1